actions: Split out zerver.actions.user_activity.
(cherry picked from commit 6168c0110a285ecbf4da9aaed19ca22dc90b925d)
