+++ zerver/lib/actions.py
@@ -45,6 +45,7 @@ from zerver.actions.default_streams import (
     get_default_streams_for_realm,
 )
 from zerver.actions.invites import notify_invites_changed, revoke_invites_generated_by_user
+from zerver.actions.user_activity import update_user_activity_interval
 from zerver.actions.user_groups import (
     do_send_user_group_members_update_event,
     update_users_in_full_members_system_group,
@@ -216,8 +217,6 @@ from zerver.models import (
     Service,
     Stream,
     Subscription,
-    UserActivity,
-    UserActivityInterval,
     UserGroup,
     UserGroupMembership,
     UserMessage,
@@ -5391,49 +5390,6 @@ def get_default_subs(user_profile: UserProfile) -> List[Stream]:
     return get_default_streams_for_realm(user_profile.realm_id)
 
 
-def do_update_user_activity_interval(
-    user_profile: UserProfile, log_time: datetime.datetime
-) -> None:
-    effective_end = log_time + UserActivityInterval.MIN_INTERVAL_LENGTH
-    # This code isn't perfect, because with various races we might end
-    # up creating two overlapping intervals, but that shouldn't happen
-    # often, and can be corrected for in post-processing
-    try:
-        last = UserActivityInterval.objects.filter(user_profile=user_profile).order_by("-end")[0]
-        # Two intervals overlap iff each interval ends after the other
-        # begins.  In this case, we just extend the old interval to
-        # include the new interval.
-        if log_time <= last.end and effective_end >= last.start:
-            last.end = max(last.end, effective_end)
-            last.start = min(last.start, log_time)
-            last.save(update_fields=["start", "end"])
-            return
-    except IndexError:
-        pass
-
-    # Otherwise, the intervals don't overlap, so we should make a new one
-    UserActivityInterval.objects.create(
-        user_profile=user_profile, start=log_time, end=effective_end
-    )
-
-
-@statsd_increment("user_activity")
-def do_update_user_activity(
-    user_profile_id: int, client_id: int, query: str, count: int, log_time: datetime.datetime
-) -> None:
-    (activity, created) = UserActivity.objects.get_or_create(
-        user_profile_id=user_profile_id,
-        client_id=client_id,
-        query=query,
-        defaults={"last_visit": log_time, "count": count},
-    )
-
-    if not created:
-        activity.count += count
-        activity.last_visit = log_time
-        activity.save(update_fields=["last_visit", "count"])
-
-
 def send_presence_changed(user_profile: UserProfile, presence: UserPresence) -> None:
     # Most presence data is sent to clients in the main presence
     # endpoint in response to the user's own presence; this results
@@ -5526,11 +5482,6 @@ def do_update_user_presence(
 
     if not user_profile.realm.presence_disabled and (created or became_online):
         send_presence_changed(user_profile, presence)
-
-
-def update_user_activity_interval(user_profile: UserProfile, log_time: datetime.datetime) -> None:
-    event = {"user_profile_id": user_profile.id, "time": datetime_to_timestamp(log_time)}
-    queue_json_publish("user_activity_interval", event)
 
 
 def update_user_presence(
