+++ zerver/worker/queue_processors.py
@@ -48,12 +48,11 @@ from zulip_bots.lib import extract_query_without_mention
 
 from zerver.actions.invites import do_send_confirmation_email
 from zerver.actions.realm_export import notify_realm_export
+from zerver.actions.user_activity import do_update_user_activity, do_update_user_activity_interval
 from zerver.context_processors import common_context
 from zerver.lib.actions import (
     do_mark_stream_messages_as_read,
     do_update_embedded_data,
-    do_update_user_activity,
-    do_update_user_activity_interval,
     do_update_user_presence,
     internal_send_private_message,
     render_incoming_message,
